backend/requirements.txt
# EcomProfit Guard — backend
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet>=3.0.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Google
google-auth==2.29.0
google-auth-oauthlib==1.3.0
google-api-python-client==2.149.0
httpx==0.28.1

# ML
numpy==2.2.1
pandas==2.2.3
statsmodels==0.14.4
catboost==1.2.10
scikit-learn==1.6.0

# Dev / тесты
pytest==8.3.4

# Utils
python-dotenv==1.0.1
python-multipart==0.0.17
openpyxl==3.1.5
pyjwt==2.10.1
